add `evict_fn` callback to `start_link`
This adds an `evict_fn` option `start_link` which accepts a function to
be called when a key/value is evicted from the cache due to it being
full. This allows cleanup / logging / instrumentation etc to be done.

The value of the key is required to execute the callback, so another
access of the ets table is required to obtain this - but this is only
done when `:evict_fn` is provided. If it's not present, behaviour is as
before, and no extra ets operations are done.

Note, this also adds an optional `opts` `Keyword` argument to `start_link`